Move LSPFileUpdates into a separate file and reduce imports.
