+++ main/lsp/LSPTask.cc
@@ -1,6 +1,8 @@
 #include "main/lsp/LSPTask.h"
 #include "absl/synchronization/notification.h"
 #include "common/sort.h"
+#include "core/NameHash.h"
+#include "core/lsp/QueryResponse.h"
 #include "main/lsp/LSPOutput.h"
 #include "main/lsp/lsp.h"
