+++ main/lsp/LSPTypechecker.cc
@@ -4,6 +4,7 @@
 #include "ast/treemap/treemap.h"
 #include "common/sort.h"
 #include "common/typecase.h"
+#include "core/ErrorQueue.h"
 #include "core/Unfreeze.h"
 #include "core/lsp/PreemptionTaskManager.h"
 #include "core/lsp/TypecheckEpochManager.h"
