+++ main/lsp/LSPTypechecker.h
@@ -2,16 +2,19 @@
 #define RUBY_TYPER_LSP_LSPTYPECHECKER_H
 
 #include "ast/ast.h"
-#include "common/concurrency/WorkerPool.h"
-#include "common/kvstore/KeyValueStore.h"
-#include "core/ErrorQueue.h"
-#include "core/NameHash.h"
 #include "core/core.h"
 #include "main/lsp/LSPConfiguration.h"
+#include "main/lsp/LSPFileUpdates.h"
+
+namespace sorbet {
+class WorkerPool;
+class KeyValueStore;
+} // namespace sorbet
 
 namespace sorbet::core::lsp {
 class PreemptionTaskManager;
-}
+class QueryResponse;
+} // namespace sorbet::core::lsp
 
 namespace sorbet::realmain::lsp {
